+++ 42/media/lua/client/Utils/BF_Lists.lua
@@ -8,7 +8,7 @@
 Function defining all of the soilable clothing.
 ]]--
 function BathroomFunctions.GetSoilableClothing()
-    local bodyLocations = {"UnderwearBottom", "Underwear", "Torso1Legs1", "Legs1", "Pants", "BathRobe", "FullSuit", "FullSuitHead", "FullTop", "BodyCostume", "ShortPants", "ShortsShort"}
+    local bodyLocations = {"UnderwearBottom", "Underwear", "Torso1Legs1", "Legs1", "Pants", "BathRobe", "FullSuit", "FullSuitHead", "FullTop", "BodyCostume", "ShortPants", "ShortsShort", "Pants_Skinny"}
     return bodyLocations
 end